restored api nesting structure; simplified init and versioning of api
