+++ pyflask/setupUtils/__init__.py
@@ -1,3 +1,2 @@
-from .configureAPI import configureAPI
 from .configureLogger import configureLogger
 from .configureRouteHandlers import configureRouteHandlers
